+++ src/main/java/com/commercetools/sync/services/impl/UnresolvedReferencesServiceImpl.java
@@ -33,7 +33,7 @@ public class UnresolvedReferencesServiceImpl implements UnresolvedReferencesServ
         "Failed to save CustomObject with key: '%s' (hash of product key: '%s').";
     private static final String DELETE_FAILED =
         "Failed to delete CustomObject with key: '%s' (hash of product key: '%s').";
-    private static final String CUSTOM_OBJECT_CONTAINER_KEY =
+    public static final String CUSTOM_OBJECT_CONTAINER_KEY =
         "commercetools-sync-java.UnresolvedReferencesService.productDrafts";